+++ Back-end/FITExamAPI/FITExamAPI/Models/Result.cs
@@ -8,7 +8,7 @@
         public int? TotalCorrectAnswer { get; set; }
         public double? Score { get; set; }
         public DateTime? ExamDate { get; set; }
-        public int? OverallTime { get; set; }
+        public double? OverallTime { get; set; }
         public virtual User? User { get; set; }
         public virtual Exam? Exam { get; set; }
         public virtual ICollection<ResultDetail>? ResultDetails { get; set; } = new List<ResultDetail>();
